% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalfdr.R
\name{empiricalfdr}
\alias{empiricalfdr}
\title{empiricalfdr}
\usage{
empiricalfdr(estimate, x, FDRLevel = 0.1)
}
\arguments{
\item{estimate}{matrix with rows containing the estimates of eta, lambda, and theta}

\item{x}{the count data in correct format}

\item{FDRLevel}{the count data in correct format}
}
\description{
calculates the false discovery rate
}
\examples{
empiricalfdr(estimate,x)
}
\keyword{cats}
