% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpEMAlgorithm.R
\name{gpEMAlgorithm}
\alias{gpEMAlgorithm}
\title{gpEMAlgorithm}
\usage{
gpEMAlgorithm(x)
}
\arguments{
\item{x}{The data.}
}
\description{
Fits a generalized Poisson by EM
}
\examples{
gpEMAlgorithm(x)
}
\keyword{cats}
