% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{protein}
\alias{protein}
\title{mutation data at speciﬁc positions within protein domains}
\format{
A data frame with 871 rows and 7 variables:
\describe{
  \item{id}{an identifier}
  \item{cd00031}{growth factors}
  \item{cd00054}{calcium-binding domain of epidermal growth factors}
  \item{cd00180}{protein kinases}
  \item{cd00204}{ankyrin domains}
  \item{cd00882}{RAS-Like GTPase family of genes}
  \item{pfam00001}{a protein domain}
}
}
\usage{
protein
}
\description{
Six functionally well-known domains to identify the hotspots of somatic variants in
TCGA sequenced tumor samples.
}
\keyword{datasets}
