% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zigpEMAlgorithm.R
\name{zigpEMAlgorithm}
\alias{zigpEMAlgorithm}
\title{zigpEMAlgorithm}
\usage{
zigpEMAlgorithm(x)
}
\arguments{
\item{x}{The data.}
}
\description{
Fits a zero-inflated generalized Poisson by EM
}
\examples{
zigpEMAlgorithm(x)
}
\keyword{cats}
