% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipEMAlgorithm.R
\name{zipEMAlgorithm}
\alias{zipEMAlgorithm}
\title{zipEMAlgorithm}
\usage{
zipEMAlgorithm(x)
}
\arguments{
\item{x}{The data.}
}
\description{
Fits a zero-inflated by EM
}
\examples{
zipEMAlgorithm(x)
}
\keyword{cats}
